requirements.txt
boto3
openai
streamlit
langchain>=0.1.17
langchain-community>=0.0.21
faiss-cpu
python-dotenv
